feat(core): show disconnected state on inspector watching badge
Extract the Vite HMR socket-state hook into a shared lib and reuse it
for the inspector "Agent is watching" pill so it flips to a red "Agent
not watching" state when the dev server drops, with a tooltip prompting
a restart.
