+++ .changeset/agent-disconnected-badge.md
@@ -2,4 +2,4 @@
 '@open-slide/core': patch
 ---
 
-Show an "Agent disconnected" state on the dev badge when the Vite HMR socket drops, with a tooltip prompting to restart the dev server.
+Flip the dev-only "Agent connected" and "Agent is watching" badges to a disconnected state when the Vite HMR socket drops, with a tooltip prompting the user to restart the dev server.
