+++ packages/core/src/locale/zh-tw.ts
@@ -201,6 +201,9 @@ export const zhTW: Locale = {
     agentWatching: 'Agent 正在關注',
     agentWatchingTooltip:
       'Agent 已經透過 dev server 看到你選的元素了，直接到聊天請它修改就行。想累積幾個再一次問才需要在這裡留 comments。',
+    agentNotWatching: 'Agent 沒在關注',
+    agentNotWatchingTooltip:
+      '已和 dev server 斷線，agent 看不到你選的元素了。請重新啟動 dev server 來恢復連線。',
     leaveComment: '留個 comment',
     commentPlaceholder: '描述你希望代理進行的修改…',
     commentShortcutHint: '⌘↵ 新增',
